requirements.txt
streamlit==0.88.0
numpy==1.19.5
tensorflow_gpu==2.6.0
Pillow==8.3.2
tensorflow==2.6.0
